Refactor the model utils functions (download weight files, get weight files path, and load models)
